Extract settings navigation to separate VM
